+++ alphadia/fdr_analysis/models/__init__.py
@@ -1,0 +1,4 @@
+from .logistic_regression import LogisticRegressionClassifier
+from .two_step_classifier import TwoStepClassifier
+
+__all__ = ["LogisticRegressionClassifier", "TwoStepClassifier"]
